refactor: 统一浮点类型 + 消除全限定名
- SensorData 传感字段 Float→Double，避免与 double 比较时的隐式拓宽精度噪声
- DeviceController 用 import HashMap 替代全限定名 java.util.HashMap

+++ src/main/java/com/example/iotplatform/model/SensorData.java
@@ -10,10 +10,10 @@ import lombok.Data;
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class SensorData {
-    private Float temp;       // 空气温度
-    private Float humi;       // 空气湿度
+    private Double temp;      // 空气温度
+    private Double humi;      // 空气湿度
     private Integer light;    // 光照度
-    private Float soilPh;     // 土壤 PH（JSON key: soilPh）
+    private Double soilPh;    // 土壤 PH（JSON key: soilPh）
     private String location;  // 可选：安装位置
     private String devType;   // 可选：设备类型
 }
